Hi, are you still using this account?
Hi!

"Madmin" is the name of a project I am working on.
Any chance you are willing to close this account so I can use the name? 
Or transfer this account over to me?

I don't know what I can offer in return.
If you have an idea, let me know.


Thanks, 
Chris.